requirements.txt
boto3==1.28.0
pyspark==3.4.1
python-dotenv==1.0.0
delta-spark==2.4.0
